requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
python-dotenv==1.0.0
openai==1.3.0
httpx==0.27.2
requests==2.31.0
python-multipart==0.0.20
pydantic==2.10.0
boto3==1.35.0

